Adding EFS share for Wordpress file and Nginx configuration

+++ user_data.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 yum update -y
-yum install -y nginx php php-fpm php-mysqlnd php-opcache php-gd php-xml php-mbstring php-curl
+yum install -y nginx php php-fpm php-mysqlnd php-opcache php-gd php-xml php-mbstring php-curl amazon-efs-utils
 
 # Install php-imagick on Amazon Linux
 yum -y install php-pear php-devel gcc ImageMagick ImageMagick-devel
@@ -13,47 +13,52 @@ yum -y install libzip libzip-devel
 # Will enable zip extension in /etc/php.ini
 pecl upgrade zip 
 
-iterator=0
-while [ $iterator -lt 10 ]
-do
-    wget https://wordpress.org/wordpress-${wp_version}.zip -O wordpress.zip
-    wget https://wordpress.org/wordpress-${wp_version}.zip.md5 -O wordpress.zip.md5
-    md5sum wordpress.zip | cut -d " " -f 1 | tr -d "\n" > sum.md5
+#Mount EFS
+mkdir /efs
+efs_id="${efs_id}"
+mount -t efs $efs_id:/ /efs
+echo $efs_id:/ /efs efs defaults,_netdev 0 0 >> /etc/fstab
 
-    if cmp -s sum.md5 wordpress.zip.md5 
-    then
-        break
-    else
-        ((iterator++))
-    fi
+if test -f -s /efs/install.lock; then
+    sleep 120
+else
 
-    sleep 20
-    if [ $iterator == '10' ] 
-    then
-        exit 1;
-    fi
-done
+    touch /efs/install.lock
+    # Install wordpress on EFS share
+    iterator=0
+    while [ $iterator -lt 10 ]
+    do
+        wget https://wordpress.org/wordpress-${wp_version}.zip -O wordpress.zip
+        wget https://wordpress.org/wordpress-${wp_version}.zip.md5 -O wordpress.zip.md5
+        md5sum wordpress.zip | cut -d " " -f 1 | tr -d "\n" > sum.md5
 
-unzip wordpress.zip -d /usr/share/nginx/html/
-mv /usr/share/nginx/html/wordpress /usr/share/nginx/html/foodbox-wordpress
-chown nginx:nginx /usr/share/nginx/html/foodbox-wordpress
-cp /usr/share/nginx/html/foodbox-wordpress/wp-config-sample.php /usr/share/nginx/html/foodbox-wordpress/wp-config.php
+        if cmp -s sum.md5 wordpress.zip.md5; then
+            break
+        else
+            ((iterator++))
+        fi
 
-# Settings php-fpm
-sed -i -e 's/user = apache/user = nginx/g' /etc/php-fpm.d/www.conf
-sed -i -e 's/group = apache/group = nginx/g' /etc/php-fpm.d/www.conf
+        sleep 20
+        if [ $iterator == '10' ]; then
+            exit 1;
+        fi
+    done
 
-# Settings Wordpress
-sed -i -e 's/database_name_here/${db_name}/g' /usr/share/nginx/html/foodbox-wordpress/wp-config.php
-sed -i -e 's/username_here/${db_user}/g' /usr/share/nginx/html/foodbox-wordpress/wp-config.php
-sed -i -e 's/password_here/${db_password}/g' /usr/share/nginx/html/foodbox-wordpress/wp-config.php
-sed -i -e 's/localhost/${db_endpoint}/g' /usr/share/nginx/html/foodbox-wordpress/wp-config.php
+    unzip wordpress.zip -d /efs/
+    mv /efs/wordpress /efs/foodbox-wordpress
+    chown nginx:nginx /efs/foodbox-wordpress
+    cp /efs/foodbox-wordpress/wp-config-sample.php /efs/foodbox-wordpress/wp-config.php
 
-# Settings nginx
-sed -i -e 's/html;/html\/foodbox-wordpress;/g' /etc/nginx/nginx.conf
+    rm -f wordpress.zip wordpress.zip.md5 sum.md5
 
-# Adding website configuration
-cat > /etc/nginx/conf.d/foodbox-wordpress.conf << EOF
+    # Settings Wordpress
+    sed -i -e 's/database_name_here/${db_name}/g' /efs/foodbox-wordpress/wp-config.php
+    sed -i -e 's/username_here/${db_user}/g' /efs/foodbox-wordpress/wp-config.php
+    sed -i -e 's/password_here/${db_password}/g' /efs/foodbox-wordpress/wp-config.php
+    sed -i -e 's/localhost/${db_endpoint}/g' /efs/foodbox-wordpress/wp-config.php
+
+    # Adding website configuration on EFS share
+    cat > /efs/foodbox-wordpress.conf <<-EOF
 server {
     listen 80;
 
@@ -89,6 +94,20 @@ server {
     }
 }
 EOF
+
+fi
+
+# Link wordpress directory and website configuration 
+ln -s /efs/foodbox-wordpress /usr/share/nginx/html/foodbox-wordpress
+ln -s /efs/foodbox-wordpress.conf /etc/nginx/conf.d/foodbox-wordpress.conf
+
+# COnfigure php-fpm
+sed -i -e 's/user = apache/user = nginx/g' /etc/php-fpm.d/www.conf
+sed -i -e 's/group = apache/group = nginx/g' /etc/php-fpm.d/www.conf
+
+# Configureterraform  nginx
+sed -i -e 's/html;/html\/foodbox-wordpress;/g' /etc/nginx/nginx.conf
+
 
 # Disable default website
 cat > /etc/nginx/nginx.conf << EOF
@@ -131,6 +150,7 @@ http {
 }
 EOF
 
+# Start and enable nginx and php-fpm
 systemctl start php-fpm
 systemctl enable php-fpm
 systemctl start nginx
